Fixing program folder
