+++ src/modules/cmdpal/src/Plugins/AllApps/Programs/LogoType.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace AllApps.Programs;
+
+public enum LogoType
+{
+    Error,
+    Colored,
+    HighContrast,
+}
